refactor(optical_flow_lk): de-duplicate the last module; jsfeatNext.ts is now a thin aggregator (#47)
Ninth and FINAL de-duplication step of #47.

- src/optical_flow_lk/optical_flow_lk.ts: replace the type-only stub with
  the REAL implementation moved verbatim from the monolith (track). The
  constructor instantiates imgproc via direct module import instead of the
  jsfeatNext.imgproc static slot.
- src/jsfeatNext.ts: now a 59-line THIN AGGREGATOR - it only imports the
  modules and attaches them to the public namespace. Unused helper imports
  (resample/convol kernels, linalg_base, fast_private, orb/yape06 utils,
  point_t, JSFEAT_CONSTANTS) pruned; they are consumed by the modules
  themselves now.

This completes the stub/monolith de-duplication: every algorithm lives in
its own real module under src/<module>/, extending the shared base from
src/core/core.ts. No type-only stubs remain; the audit's section-4 target
architecture (thin aggregator) is reached for the module layer.

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57 (the
optical_flow_lk parity test pins track() vs original jsfeat across pyramid
levels); UMD bundle checked - all 18 public modules attach and instanceof
chains hold.
